refract

+++ lib/dashboard/models/home_data_model.dart
@@ -112,7 +112,7 @@ class MyStore {
   String? storeCategoryName;
   String? zone;
   String? deliveryType;
-  int? deliveryFee;
+  double? deliveryFee;
   String? hubUuid;
   bool? isApproved;
   bool? isDeleted;
@@ -180,7 +180,7 @@ class MyStore {
         storeCategoryName: json["storeCategoryName"],
         zone: json["zone"],
         deliveryType: json["deliveryType"],
-        deliveryFee: json["deliveryFee"],
+        deliveryFee: json["deliveryFee"].toDouble(),
         hubUuid: json["hubUuid"],
         isApproved: json["isApproved"],
         isDeleted: json["isDeleted"],
