+++ lib/dashboard/orders/models/order_response_model.dart
@@ -83,7 +83,7 @@ class Docs {
   dynamic operatorUuid;
   bool? isActive;
   List<ProductDetail>? productDetails;
-  int? storeDeliverycharge;
+  double? storeDeliverycharge;
   List<dynamic>? additionalChargesArray;
   int? v;
 
@@ -146,7 +146,7 @@ class Docs {
             ? []
             : List<ProductDetail>.from(
                 json["productDetails"]!.map((x) => ProductDetail.fromJson(x))),
-        storeDeliverycharge: json["storeDeliverycharge"],
+        storeDeliverycharge: json["storeDeliverycharge"].toDouble(),
         additionalChargesArray: json["additionalChargesArray"] == null
             ? []
             : List<dynamic>.from(json["additionalChargesArray"]!.map((x) => x)),
